+++ data/service/api/v1/datasets_update.go
@@ -10,7 +10,7 @@ import (
 )
 
 func DatasetsUpdate(dataServiceContext dataService.Context) {
-	datasetID := dataServiceContext.Request().PathParam("datasetid")
+	datasetID := dataServiceContext.Request().PathParam("dataset_id")
 	if datasetID == "" {
 		dataServiceContext.RespondWithError(ErrorDatasetIDMissing())
 		return