+++ data/service/api/v1/users_data_delete.go
@@ -8,7 +8,7 @@ import (
 )
 
 func UsersDataDelete(dataServiceContext dataService.Context) {
-	targetUserID := dataServiceContext.Request().PathParam("userid")
+	targetUserID := dataServiceContext.Request().PathParam("user_id")
 	if targetUserID == "" {
 		dataServiceContext.RespondWithError(ErrorUserIDMissing())
 		return