+++ data/service/api/v1/users_datasets_get.go
@@ -18,7 +18,7 @@ const (
 )
 
 func UsersDatasetsGet(dataServiceContext dataService.Context) {
-	targetUserID := dataServiceContext.Request().PathParam("userid")
+	targetUserID := dataServiceContext.Request().PathParam("user_id")
 	if targetUserID == "" {
 		dataServiceContext.RespondWithError(ErrorUserIDMissing())
 		return